requirements.txt
absl-py==0.12.0
anyio==3.1.0
apturl
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fastapi==0.63.0
fasteners==0.14.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.0.4
h11==0.12.0
h2==4.0.0
h5py==3.1.0
hpack==4.0.0
httplib2==0.14.0
httptools==0.1.1
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.8
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.2
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
priority==1.3.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydantic==1.8.2
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
pytest==6.2.2
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
starlette==0.13.6
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-cpu==2.4.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tk==0.1.0
tkinter-page==0.0.5
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.13.3
uvloop==0.14.0
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.1
winregistry==0.8.3
wrapt==1.12.1
wsproto==1.0.0
xkit==0.0.0
